// frontend/src/client/models/UserCreate.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type UserCreate = {
    username?: string;
    hashed_password?: string;
    email?: string;
    mobile?: string;
    user_photo?: string;
    first_name?: string;
    last_name?: string;
    gender?: string;
    date_of_birth?: string;
    mobile_validation?: boolean;
    email_validation?: boolean;
    disabled?: boolean;
    is_admin?: boolean;
    is_superuser?: boolean;
    player_id?: number;
    coach_id?: number;
    referee_id?: number;
    manager_id?: number;
    password: string;
};